refactor: move projection pushing to it's own file
